refactor(api): use directly mocks/attestation-storage.mock.js in tests
