+++ api/tests/prescription/organization-learner/integration/domain/usecases/find-paginated-filtered-attestation-participants-status_test.js
@@ -1,6 +1,7 @@
 import { AttestationParticipantStatus } from '../../../../../../src/prescription/organization-learner/domain/read-models/AttestationParticipantStatus.js';
 import { usecases } from '../../../../../../src/prescription/organization-learner/domain/usecases/index.js';
-import { databaseBuilder, expect, mockAttestationStorage } from '../../../../../test-helper.js';
+import { databaseBuilder, expect } from '../../../../../test-helper.js';
+import { mockAttestationStorage } from '../../../../../tooling/mocks/attestation-storage.mock.js';
 
 describe('Integration | Prescription | Learner Management | Domain | UseCase | find-paginated-filtered-attestation-participant-status', function () {
   let attestation, organizationId, firstLearner, secondLearner;
