+++ api/tests/profile/acceptance/application/attestation-route_test.js
@@ -3,8 +3,8 @@ import {
   databaseBuilder,
   expect,
   generateAuthenticatedUserRequestHeaders,
-  mockAttestationStorage,
 } from '../../../test-helper.js';
+import { mockAttestationStorage } from '../../../tooling/mocks/attestation-storage.mock.js';
 
 describe('Profile | Acceptance | Application | Attestation Route ', function () {
   let server;
